+++ pypose/__init__.py
@@ -13,10 +13,11 @@ from .lietensor import Sim3_type, sim3_type, RxSO3_type, rxso3_type
 from .lietensor import tensor, translation, rotation, scale, matrix, euler
 from .lietensor import mat2SO3, mat2SE3, mat2Sim3, mat2RxSO3, from_matrix, matrix, euler2SO3, vec2skew
 from .lietensor import gradcheck, gradgradcheck
-from .lietensor.function import *
-from . import module, optim
+# from .lietensor.function import *
 from .function import *
 from .basics import *
+from . import module
+from . import optim
 
 
 min_torch = '2.0'
