+++ pypose/function/__init__.py
@@ -1,2 +1,3 @@
 # functions implemented in this module need the lietensor module
+from .linalg import *
 from .geometry import *
